Refactor service functionality to access kardianos/service.Service interface
